backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
celery-redbeat==2.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-ninja==1.3.0
django-redis==5.4.0
flower==2.0.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
humanize==4.12.1
idna==3.10
kombu==5.4.2
oauth2client==4.1.3
oauthlib==3.2.2
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tenacity==9.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
